Rename option numbers
