+++ examples/example-otel/cmd/client/main.go
@@ -8,8 +8,8 @@ import (
 	"go.opentelemetry.io/otel"
 	"google.golang.org/grpc"
 
-	echov1 "github.com/clly/proto-telemetry/example-otel/gen/proto/go/otecho/v1"
-	"github.com/clly/proto-telemetry/example-otel/tracing"
+	"github.com/clly/proto-telemetry/examples/example-otel/gen/proto/go/otecho/v1"
+	"github.com/clly/proto-telemetry/examples/example-otel/tracing"
 )
 
 func main() {
@@ -32,17 +32,17 @@ func run() error {
 	}
 	defer shutdown()
 
-	echo := echov1.NewEchoServiceClient(conn)
+	echo := otechov1.NewEchoServiceClient(conn)
 
 	ctx := context.Background()
 	ctx, span := otel.Tracer("protoc-gen-go-telemetry/example/client").Start(ctx, "Echo Client")
 	defer span.End()
 
-	req := &echov1.EchoRequest{
+	req := &otechov1.EchoRequest{
 		Msg: "Hello World!",
 	}
 	req.TraceAttributes(ctx)
-	if _, err := echo.Echo(context.Background(), &echov1.EchoRequest{
+	if _, err := echo.Echo(context.Background(), &otechov1.EchoRequest{
 		Msg: "Hello World!",
 	}); err != nil {
 		return fmt.Errorf("failed to Echo: %w", err)
