+++ proto/telemetry/options/v1/options.proto
@@ -6,20 +6,18 @@ import "google/protobuf/descriptor.proto";
 
 option go_package = "github.com/clly/proto-telemetry/proto/telemetry/options/v1;v1";
 
+// global number: 1179
+
 extend google.protobuf.FileOptions {
-  string telemetry_package = 10000;
+  string exclude_file = 11790;
 }
 
 extend google.protobuf.FieldOptions {
-  bool exclude = 10000;
-  string field_name = 10001;
+  bool exclude = 11791;
+  string field_name = 11792;
 }
 
 extend google.protobuf.MessageOptions {
-  bool exclude_message = 10010;
-  string message_name = 10011;
-}
-
-extend google.protobuf.FileOptions {
-  bool exclude_file = 10020;
+  bool exclude_message = 11793;
+  string message_name = 11794;
 }
