+++ run-dev.sh
@@ -6,13 +6,13 @@ go install ./cmd/protoc-gen-go-telemetry/
 
 (
     echo "Generating opentelemetry"
-    cd example-otel
+    cd examples/example-otel
     buf generate
 )
 
 echo 
 (
     echo "Generating opencensus"
-    cd example-oc
+    cd examples/example-oc
     buf generate
 )
